+++ packages/contracts/hardhat.config.ts
@@ -4,6 +4,7 @@ import "@nomiclabs/hardhat-waffle";
 import "hardhat-gas-reporter";
 import "hardhat-deploy";
 import "solidity-coverage";
+import "@tenderly/hardhat-tenderly";
 
 import { config as dotEnvConfig } from "dotenv";
 import { HardhatUserConfig } from "hardhat/types";
@@ -17,6 +18,8 @@ import "./src/tasks/getChainId";
 import "./src/tasks/renounceOwnership";
 import "./src/tasks/proposeTransferOwnership";
 import "./src/tasks/setAggregator";
+import "./src/tasks/setDexPrice";
+import "./src/tasks/decodeInputData";
 
 dotEnvConfig();
 
@@ -73,7 +76,7 @@ const config: HardhatUserConfig = {
     mainnet: {
       accounts: { mnemonic },
       chainId: 1,
-      url: urlOverride || "http://localhost:8545",
+      url: urlOverride || process.env.ETH_PROVIDER_URL || "https://cloudflare-eth.com",
     },
     ropsten: {
       accounts: { mnemonic },
@@ -108,7 +111,7 @@ const config: HardhatUserConfig = {
     xdai: {
       accounts: { mnemonic },
       chainId: 100,
-      url: urlOverride || process.env.XDAI_PROVIDER_URL || "https://rpc.xdaichain.com/",
+      url: urlOverride || process.env.XDAI_PROVIDER_URL || "https://xdai.poanetwork.dev/",
     },
     matic: {
       accounts: { mnemonic },