+++ packages/integration/test/load/routerCyclical.ts
@@ -1,6 +1,7 @@
 import { delay, Logger } from "@connext/nxtp-utils";
 import { utils } from "ethers";
 import pino from "pino";
+import { getDecimals } from "@connext/nxtp-sdk/src/utils";
 
 import { getConfig } from "../utils/config";
 import { SdkManager } from "../utils/sdkManager";
@@ -14,7 +15,8 @@ import { writeStatsToFile } from "../utils/reporting";
  */
 const routerCyclical = async (numberOfAgents: number, duration: number) => {
   const config = getConfig();
-  const log = pino({ level: config.logLevel ?? "info" });
+  const log = pino({ level: "error" });
+  const amount = utils.parseEther("100").toString();
 
   const durationMs = duration * 60 * 1000;
   // Create manager
@@ -25,13 +27,14 @@ const routerCyclical = async (numberOfAgents: number, duration: number) => {
     new Logger({ level: config.logLevel ?? "info" }),
     config.natsUrl,
     config.authUrl,
+    // config.network
   );
   log.info({ agents: numberOfAgents }, "Created manager");
 
   // Get transfer config
   const sendingChainId = parseInt(Object.keys(config.chainConfig)[0]);
   const receivingChainId = parseInt(Object.keys(config.chainConfig)[1]);
-  log.info({ sendingChainId, receivingChainId }, "Chose chains");
+  log.info({ sendingChainId, receivingChainId }, "Picked chains");
   const swap = config.swapPools.find((swap) => {
     // Must have sending and receiving chain
     const chains = swap.assets.map((a) => a.chainId);
@@ -44,43 +47,51 @@ const routerCyclical = async (numberOfAgents: number, duration: number) => {
   const { assetId: receivingAssetId } = swap.assets.find((a) => a.chainId === receivingChainId)!;
 
   // Fund agents with tokens on sending + receiving chain
-  await manager.giftAgentsOnchain(sendingAssetId, sendingChainId);
+  if (manager) {
+    log.info(`Gifting agents sending chain ${sendingChainId}`);
+    await manager.giftAgentsOnchain(sendingAssetId, sendingChainId);
 
-  await manager.giftAgentsOnchain(receivingAssetId, receivingChainId);
+    log.info(`Gifting agents receiving chain ${receivingChainId}`);
+    await manager.giftAgentsOnchain(receivingAssetId, receivingChainId);
 
-  // Begin transfers
-  log.warn({ duration, numberOfAgents }, "Beginning cyclical test");
+    // Begin transfers
+    log.warn({ duration, numberOfAgents }, "Beginning cyclical test");
 
-  const startTime = Date.now();
-  const killSwitch = await manager.startCyclicalTransfers({
-    sendingAssetId,
-    sendingChainId,
-    receivingAssetId,
-    receivingChainId,
-    amount: utils.parseEther("0.0000001").toString(),
-  });
+    const provider = config.chainConfig[sendingChainId].provider;
+    const decimals = await getDecimals(sendingAssetId, provider);
+    const amount = utils.parseUnits("0.0001", decimals).toString();
 
-  await new Promise((resolve) => {
-    setTimeout(() => {
-      log.warn({ duration, numberOfAgents, durationMs, startTime, now: Date.now() }, "Activating kill switch");
-      killSwitch();
-      resolve(undefined);
-    }, durationMs);
-  });
+    const startTime = Date.now();
+    const killSwitch = await manager.startCyclicalTransfers({
+      sendingAssetId,
+      sendingChainId,
+      receivingAssetId,
+      receivingChainId,
+      amount,
+    });
 
-  // Wait 90s for stragglers
-  await delay(90 * 1000);
+    await new Promise((resolve) => {
+      setTimeout(() => {
+        log.warn({ duration, numberOfAgents, durationMs, startTime, now: Date.now() }, "Activating kill switch");
+        killSwitch();
+        resolve(undefined);
+      }, durationMs);
+    });
 
-  log.warn({ duration, numberOfAgents }, "Test complete, printing summary");
+    // Wait 90s for stragglers
+    await delay(90 * 1000);
 
-  const summary = manager.getTransferSummary();
-  log.error(summary, "Transfer summary");
+    log.warn({ duration, numberOfAgents }, "Test complete, printing summary");
 
-  // Write transfer summary to file
-  writeStatsToFile(`router.cyclical`, summary);
+    const summary = manager.getTransferSummary();
+    log.error(summary, "Transfer summary");
 
-  log.error("Test complete");
-  process.exit(0);
+    // Write transfer summary to file
+    writeStatsToFile(`router.cyclical`, summary);
+
+    log.error("Test complete");
+    process.exit(0);
+  }
 };
 
-routerCyclical(parseInt(process.env.NUMBER_OF_AGENTS ?? "5"), parseInt(process.env.DURATION ?? "20"));
+routerCyclical(parseInt(process.env.NUMBER_OF_AGENTS ?? "1"), parseInt(process.env.DURATION ?? "10"));