+++ packages/router/src/lib/entities/contractReader.ts
@@ -62,5 +62,3 @@ export type SingleChainTransaction = {
   signature?: string; // only there when fulfilled or cancelled
   relayerFee?: string; // only there when fulfilled or cancelled
 };
-
-export type SubgraphSyncRecord = { synced: boolean; latestBlock: number; syncedBlock: number };