+++ packages/sdk/src/types.ts
@@ -4,28 +4,90 @@ import {
   TIntegerString,
   TransactionDataSchema,
   PrepareParams,
+  Logger,
   TransactionPreparedEvent,
   TransactionFulfilledEvent,
   TransactionCancelledEvent,
   CrosschainTransaction,
+  UserNxtpNatsMessagingService,
 } from "@connext/nxtp-utils";
 import { Type, Static } from "@sinclair/typebox";
-import { providers } from "ethers";
+import { providers, Signer } from "ethers";
+
+export const SdkBaseChainConfigSchema = Type.Record(
+  Type.Number(),
+  Type.Object({
+    provider: Type.Any(providers.FallbackProvider),
+    transactionManagerAddress: Type.Optional(Type.String()),
+    priceOracleAddress: Type.Optional(Type.String()),
+    subgraph: Type.Optional(Type.String()),
+    subgraphSyncBuffer: Type.Optional(Type.Number()),
+  }),
+);
+
+export const LogLevelScehma = Type.Union([
+  Type.Literal("fatal"),
+  Type.Literal("error"),
+  Type.Literal("warn"),
+  Type.Literal("info"),
+  Type.Literal("debug"),
+  Type.Literal("trace"),
+  Type.Literal("silent"),
+]);
+
+export const NetworkSchema = Type.Union([Type.Literal("local"), Type.Literal("testnet"), Type.Literal("mainnet")]);
+
+// export const SdkBaseConfigSchema = Type.Object({
+//   chainConfig: SdkBaseChainConfigParams,
+//   signerAddress: Type.Promise(Type.String()),
+//   signer: Type.Optional(Type.Any(Signer)),
+//   messagingSigner: Type.Optional(Type.Any(Signer)),
+//   logger: Type.Optional(Type.Any(Logger)),
+//   network: Type.Optional(Type.Enum(NetworkEnum)),
+//   natsUrl: Type.Optional(Type.String()),
+//   authUrl: Type.Optional(Type.String()),
+//   messaging: Type.Optional(Type.Any(UserNxtpNatsMessagingService)),
+//   skipPolling: Type.Optional(Type.Boolean()),
+// });
+
+// export type SdkBaseConfigParams = Static<typeof SdkBaseConfigSchema>;
+export type SdkBaseChainConfigParams = {
+  [chainId: number]: {
+    provider: providers.FallbackProvider;
+    transactionManagerAddress?: string;
+    priceOracleAddress?: string;
+    subgraph?: string | string[];
+    subgraphSyncBuffer?: number;
+  };
+};
+export type SdkBaseConfigParams = {
+  chainConfig: SdkBaseChainConfigParams;
+  signerAddress: Promise<string>;
+  signer?: Signer;
+  messagingSigner?: Signer;
+  logger?: Logger;
+  network?: "testnet" | "mainnet" | "local";
+  natsUrl?: string;
+  authUrl?: string;
+  messaging?: UserNxtpNatsMessagingService;
+  skipPolling?: boolean;
+};
 
 export const CrossChainParamsSchema = Type.Object({
-  callData: Type.Optional(Type.RegEx(/^0x[a-fA-F0-9]*$/)),
   sendingChainId: TChainId,
   sendingAssetId: TAddress,
   receivingChainId: TChainId,
   receivingAssetId: TAddress,
-  callTo: Type.Optional(TAddress),
   receivingAddress: TAddress,
   amount: TIntegerString,
+  callTo: Type.Optional(TAddress),
+  callData: Type.Optional(Type.RegEx(/^0x[a-fA-F0-9]*$/)),
+  encryptedCallData: Type.Optional(Type.String()),
   expiry: Type.Optional(Type.Number()),
   transactionId: Type.Optional(Type.RegEx(/^0x[a-fA-F0-9]{64}$/)),
   slippageTolerance: Type.Optional(Type.String()),
   dryRun: Type.Optional(Type.Boolean()),
-  preferredRouter: Type.Optional(TAddress),
+  preferredRouters: Type.Optional(Type.Array(TAddress)),
   initiator: Type.Optional(TAddress),
 });
 
@@ -53,15 +115,6 @@ export const AuctionBidParamsSchema = Type.Object({
 });
 
 export type AuctionBidParams = Static<typeof AuctionBidParamsSchema>;
-
-export const TransactionPrepareEventSchema = Type.Object({
-  txData: TransactionDataSchema,
-  encryptedCallData: Type.String(),
-  encodedBid: Type.String(),
-  bidSignature: Type.String(),
-});
-
-export type TransactionPrepareEventParams = Static<typeof TransactionPrepareEventSchema>;
 
 export const CancelSchema = Type.Object({
   txData: TransactionDataSchema,