+++ packages/sdk/test/integration/transactionManager.spec.ts
@@ -10,15 +10,10 @@ import {
   expect,
   Logger,
 } from "@connext/nxtp-utils";
-import { utils, constants, BigNumber } from "ethers";
+import { utils, constants } from "ethers";
 
-import {
-  Counter,
-  TransactionManager as TransactionManagerTypechain,
-  TestERC20,
-} from "@connext/nxtp-contracts/typechain";
+import { TransactionManager as TransactionManagerTypechain, TestERC20 } from "@connext/nxtp-contracts/typechain";
 import TransactionManagerArtifact from "@connext/nxtp-contracts/artifacts/contracts/TransactionManager.sol/TransactionManager.json";
-import CounterArtifact from "@connext/nxtp-contracts/artifacts/contracts/test/Counter.sol/Counter.json";
 import TestERC20Artifact from "@connext/nxtp-contracts/artifacts/contracts/test/TestERC20.sol/TestERC20.json";
 
 import { approveTokens, addPrivileges, prepareAndAssert } from "../helper";
@@ -28,7 +23,7 @@ import {
   getDeployedPriceOracleContract,
   getDeployedChainIdsForGasFee,
 } from "../../src/transactionManager/transactionManager";
-import { ChainNotConfigured, PriceOracleNotConfigured } from "../../src/error";
+import { ChainNotConfigured } from "../../src/error";
 import { deployContract } from "../../../contracts/test/utils";
 
 const { AddressZero } = constants;
@@ -44,7 +39,7 @@ const createFixtureLoader = waffle.createFixtureLoader;
 describe("Transaction Manager", function () {
   const [wallet, router, user, receiver] = waffle.provider.getWallets();
 
-  const sendingChainId = 1337;
+  const sendingChainId = 31337;
   const receivingChainId = 1338;
   const routerFunds = "1000";
   const userFunds = "100";
@@ -55,7 +50,6 @@ describe("Transaction Manager", function () {
   let routerTransactionManager: TransactionManager;
   let transactionManager: TransactionManagerTypechain;
   let transactionManagerReceiverSide: TransactionManagerTypechain;
-  let counter: Counter;
   let tokenA: TestERC20;
   let tokenB: TestERC20;
 
@@ -99,8 +93,6 @@ describe("Transaction Manager", function () {
       receivingChainId,
     );
 
-    counter = await deployContract<Counter>(CounterArtifact);
-
     tokenA = await deployContract<TestERC20>(TestERC20Artifact);
 
     tokenB = await deployContract<TestERC20>(TestERC20Artifact);
@@ -140,10 +132,12 @@ describe("Transaction Manager", function () {
         [sendingChainId]: {
           provider: user.provider,
           transactionManagerAddress: transactionManager.address,
+          priceOracleAddress: undefined,
         },
         [receivingChainId]: {
           provider: user.provider,
           transactionManagerAddress: transactionManagerReceiverSide.address,
+          priceOracleAddress: undefined,
         },
       },
       user.getAddress(),
@@ -155,10 +149,12 @@ describe("Transaction Manager", function () {
         [sendingChainId]: {
           provider: router.provider,
           transactionManagerAddress: transactionManager.address,
+          priceOracleAddress: undefined,
         },
         [receivingChainId]: {
           provider: router.provider,
           transactionManagerAddress: transactionManagerReceiverSide.address,
+          priceOracleAddress: undefined,
         },
       },
       router.getAddress(),
@@ -173,56 +169,34 @@ describe("Transaction Manager", function () {
   });
 
   describe("#getDeployedTransactionManagerContract", () => {
-    it("should undefined if no transaction manager", () => {
-      const res = getDeployedTransactionManagerContract(0);
-      expect(res).to.be.undefined;
-    });
-
-    it("happy func", () => {
-      const res = getDeployedTransactionManagerContract(4);
-      expect(res).to.be.ok;
-    });
-  });
-
-  describe("getDeployedTransactionManagerContract", () => {
-    it("happy case: returns undefined", async () => {
+    it("should return undefined if no transaction manager", async () => {
       const chainId = sendingChainId;
       const res = getDeployedTransactionManagerContract(chainId);
       expect(res).to.be.undefined;
     });
-    it("happy case", async () => {
-      const chainId = 5;
+
+    it("happy", async () => {
+      const chainId = 4;
       const res = getDeployedTransactionManagerContract(chainId);
       expect(res.address).to.be.a("string");
     });
   });
 
   describe("#getDeployedPriceOracleContract", () => {
-    it("should undefined if no transaction manager", () => {
+    it("should return undefined if no price oracle contract", () => {
       const res = getDeployedPriceOracleContract(0);
       expect(res).to.be.undefined;
     });
 
-    it("happy func", () => {
-      const res = getDeployedPriceOracleContract(4);
-      expect(res).to.be.ok;
-    });
-  });
-
-  describe("getDeployedPriceOracleContract", () => {
-    it("happy case: returns undefined", async () => {
-      const chainId = sendingChainId;
-      const res = getDeployedPriceOracleContract(chainId);
-      expect(res).to.be.undefined;
-    });
-    it("happy case", async () => {
+    it("happy", () => {
       const chainId = 4;
       const res = getDeployedPriceOracleContract(chainId);
+      expect(res).to.be.ok;
       expect(res.address).to.be.a("string");
     });
   });
 
-  describe("getDeployedChainIdsForGasFee", () => {
+  describe("#getDeployedChainIdsForGasFee", () => {
     it("happy case", async () => {
       const res = getDeployedChainIdsForGasFee();
       expect(res).to.be.includes(4);
@@ -244,7 +218,7 @@ describe("Transaction Manager", function () {
       );
     });
 
-    describe("prepare", () => {
+    describe("#prepare", () => {
       it("should error if unfamiliar chainId", async () => {
         const { transaction, record } = await getTransactionData();
         const InvalidChainId = 123;
@@ -270,7 +244,7 @@ describe("Transaction Manager", function () {
       });
     });
 
-    describe("cancel", () => {
+    describe("#cancel", () => {
       const relayerFee = "1";
 
       it("should error if unfamiliar chainId", async () => {
@@ -341,7 +315,7 @@ describe("Transaction Manager", function () {
       });
     });
 
-    describe("fulfill", () => {
+    describe("#fulfill", () => {
       const relayerFee = "1";
 
       it("should error if unfamiliar chainId", async () => {
@@ -406,12 +380,18 @@ describe("Transaction Manager", function () {
       });
     });
 
-    describe("approveTokensIfNeeded", () => {
+    describe("#approveTokensIfNeeded", () => {
       it("should error if unfamiliar chainId", async () => {
         const InvalidChainId = 123;
         await expect(
           userTransactionManager.approveTokensIfNeeded(InvalidChainId, tokenA.address, "1"),
         ).to.be.rejectedWith(ChainNotConfigured.getMessage(InvalidChainId, supportedChains));
+      });
+
+      it("happy case: when allowance is suffice return undefined", async () => {
+        const approveReq = await userTransactionManager.approveTokensIfNeeded(sendingChainId, tokenA.address, "0");
+
+        expect(approveReq).to.be.undefined;
       });
 
       it("happy case", async () => {
@@ -423,15 +403,15 @@ describe("Transaction Manager", function () {
       });
     });
 
-    describe.skip("establishListeners", () => {
+    describe.skip("#establishListeners", () => {
       it("happy case", async () => {});
     });
 
-    describe.skip("removeAllListeners", () => {
+    describe.skip("#removeAllListeners", () => {
       it.skip("happy case", async () => {});
     });
 
-    describe("getRouterLiquidity", () => {
+    describe("#getRouterLiquidity", () => {
       it("should error if unfamiliar chainId", async () => {
         const InvalidChainId = 123;
         await expect(
@@ -451,7 +431,7 @@ describe("Transaction Manager", function () {
       });
     });
 
-    describe("calculateGasInTokenForFullfill", () => {
+    describe("#calculateGasInTokenForFullfill", () => {
       it("should error if unfamiliar chainId", async () => {
         const InvalidChainId = 123;
         const { transaction, record } = await getTransactionData();