+++ packages/sdk/test/unit/sdkBase.spec.ts
@@ -7,6 +7,7 @@ import {
   VariantTransactionData,
   AuctionBid,
   Logger,
+  delay,
 } from "@connext/nxtp-utils";
 import { expect } from "chai";
 import { providers, Wallet, constants, BigNumber } from "ethers";
@@ -44,6 +45,10 @@ const logger = new Logger({ level: process.env.LOG_LEVEL ?? "silent" });
 const { AddressZero } = constants;
 const response = "connected";
 
+// NOTE: Tried importing the evt module Timeout error here, but the compiler throws.
+// For now, just mocking using the proper message string.
+const mockEvtTimeoutErr = (timeout: number) => new Error(`Evt timeout after ${timeout}ms`);
+
 describe("NxtpSdkBase", () => {
   let sdk: NxtpSdkBase;
   let signer: SinonStubbedInstance<Wallet>;
@@ -54,8 +59,6 @@ describe("NxtpSdkBase", () => {
   let provider1338: SinonStubbedInstance<providers.FallbackProvider>;
   let signFulfillTransactionPayloadMock: SinonStub;
   let recoverAuctionBidMock: SinonStub;
-  let ethereumRequestMock: SinonStub;
-  let encryptMock: SinonStub;
   let balanceStub: SinonStub;
 
   let user: string = getAddress(mkAddress("0xa"));
@@ -101,8 +104,6 @@ describe("NxtpSdkBase", () => {
 
     signFulfillTransactionPayloadMock = stub(utils, "signFulfillTransactionPayload");
     recoverAuctionBidMock = stub(utils, "recoverAuctionBid");
-    ethereumRequestMock = stub(utils, "ethereumRequest");
-    encryptMock = stub(utils, "encrypt");
     recoverAuctionBidMock.returns(router);
 
     stub(sdkIndex, "AUCTION_TIMEOUT").value(1_000);
@@ -183,6 +184,7 @@ describe("NxtpSdkBase", () => {
     const transactionId = getRandomBytes32();
     const crossChainParams = {
       callData: EmptyBytes,
+      encryptedCallData: EmptyBytes,
       sendingChainId: sendingChainId,
       sendingAssetId: mkAddress("0xc"),
       receivingChainId: receivingChainId,
@@ -306,19 +308,10 @@ describe("NxtpSdkBase", () => {
       expect(messaging.connect.callCount).to.be.eq(1);
     });
 
-    it("should fail if subscribeToMetaTxResponse fails", async () => {
-      messaging.subscribeToMetaTxResponse.rejects(new Error("fail"));
-
-      await expect(sdk.connectMessaging()).to.be.rejectedWith("fail");
-      expect(messaging.connect.callCount).to.be.eq(1);
-      expect(messaging.subscribeToAuctionResponse.callCount).to.be.eq(1);
-    });
-
     it("should work", async () => {
       await sdk.connectMessaging();
       expect(messaging.connect.callCount).to.be.eq(1);
       expect(messaging.subscribeToAuctionResponse.callCount).to.be.eq(1);
-      expect(messaging.subscribeToMetaTxResponse.callCount).to.be.eq(1);
     });
   });
 
@@ -427,28 +420,6 @@ describe("NxtpSdkBase", () => {
       );
     });
 
-    it("should error if eth_getEncryptionPublicKey errors", async () => {
-      const callData = getRandomBytes32();
-      const { crossChainParams } = getMock({ callData });
-
-      ethereumRequestMock.throws(new Error("fails"));
-
-      await expect(sdk.getTransferQuote(crossChainParams)).to.be.rejectedWith(
-        EncryptionError.getMessage("public key encryption failed"),
-      );
-    });
-
-    it("should fail if encrypt fails", async () => {
-      const callData = getRandomBytes32();
-      const { crossChainParams } = getMock({ callData });
-
-      encryptMock.throws(new Error("fails"));
-
-      await expect(sdk.getTransferQuote(crossChainParams)).to.be.rejectedWith(
-        EncryptionError.getMessage("public key encryption failed"),
-      );
-    });
-
     it("should not include improperly signed bids", async () => {
       const { crossChainParams, auctionBid, bidSignature } = getMock();
 
@@ -546,13 +517,14 @@ describe("NxtpSdkBase", () => {
 
     it("happy: should get a transfer quote with callData", async () => {
       const callData = getRandomBytes32();
-      const randomHash = keccak256(getRandomBytes32());
-      const { crossChainParams, auctionBid, bidSignature } = getMock({ callData });
+      const encryptedCallData = keccak256(getRandomBytes32());
+      const { crossChainParams, auctionBid, bidSignature } = getMock({
+        callData,
+        encryptedCallData,
+      });
 
       recoverAuctionBidMock.returns(auctionBid.router);
       transactionManager.getRouterLiquidity.resolves(BigNumber.from(auctionBid.amountReceived));
-      ethereumRequestMock.resolves(randomHash);
-      encryptMock.resolves(randomHash);
 
       setTimeout(() => {
         messageEvt.post({ inbox: "inbox", data: { bidSignature, bid: auctionBid, gasFeeInReceivingToken: "0" } });
@@ -582,7 +554,7 @@ describe("NxtpSdkBase", () => {
       setTimeout(() => {
         messageEvt.post({ inbox: "inbox", data: { bidSignature, bid: auctionBid, gasFeeInReceivingToken: "0" } });
       }, 150);
-      const res = await sdk.getTransferQuote({ ...crossChainParams, preferredRouter: auctionBid.router });
+      const res = await sdk.getTransferQuote({ ...crossChainParams, preferredRouters: [auctionBid.router] });
 
       expect(res.bid).to.be.eq(auctionBid);
       expect(res.bidSignature).to.be.eq(bidSignature);
@@ -714,7 +686,7 @@ describe("NxtpSdkBase", () => {
     });
 
     describe("should error if invalid config", () => {
-      it("unkown sendingChainId", async () => {
+      it("unknown sendingChainId", async () => {
         const { transaction, record } = await getTransactionData({ sendingChainId: 1400 });
         await expect(
           sdk.fulfillTransfer(
@@ -733,7 +705,7 @@ describe("NxtpSdkBase", () => {
         ).to.eventually.be.rejectedWith(ChainNotConfigured.getMessage(1400, supportedChains));
       });
 
-      it("unkown receivingChainId", async () => {
+      it("unknown receivingChainId", async () => {
         const { transaction, record } = await getTransactionData({ receivingChainId: 1400 });
 
         await expect(
@@ -756,14 +728,9 @@ describe("NxtpSdkBase", () => {
 
     it("should error if finish transfer => useRelayers:true, metaTxResponse errors", async () => {
       const { transaction, record } = await getTransactionData();
-      stub(sdkIndex, "META_TX_TIMEOUT").value(1_000);
+      stub(sdkIndex, "META_TX_TIMEOUT").value(100);
 
-      setTimeout(() => {
-        messageEvt.post({
-          inbox: "inbox",
-          err: "Blahhh",
-        });
-      }, 200);
+      subgraph.waitFor.rejects(mockEvtTimeoutErr(100));
 
       await expect(
         sdk.fulfillTransfer(
@@ -779,7 +746,7 @@ describe("NxtpSdkBase", () => {
           "0",
           true,
         ),
-      ).to.be.rejectedWith(MetaTxTimeout.getMessage(1_000));
+      ).to.be.rejectedWith(MetaTxTimeout.getMessage(100));
     });
 
     it("happy: finish transfer => useRelayers:true", async () => {
@@ -787,15 +754,15 @@ describe("NxtpSdkBase", () => {
 
       const transactionHash = mkHash("0xc");
 
-      setTimeout(() => {
-        messageEvt.post({
-          inbox: "inbox",
-          data: {
-            transactionHash,
-            chainId: sendingChainId,
-          },
-        });
-      }, 200);
+      subgraph.waitFor.resolves({
+        transactionHash,
+        txData: {
+          ...transaction,
+          ...record,
+          sendingChainId: receivingChainId,
+          receivingChainId: sendingChainId,
+        },
+      } as any);
 
       const res = await sdk.fulfillTransfer(
         {